Change Debugger.Config to follow convention of other classes
