+++ src/config.ts
@@ -1,0 +1,57 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import { murmur2 } from 'murmurhash-js';
+
+import { IDebugger } from './tokens';
+
+/**
+ * A class that holds debugger configuration for all kernels.
+ */
+export class DebuggerConfig implements IDebugger.IConfig {
+  /**
+   * Returns an id based on the given code.
+   *
+   * @param code The source code.
+   * @param kernel The kernel name from current session.
+   */
+  getCodeId(code: string, kernel: string): string {
+    const { prefix, suffix } = this._fileParams.get(kernel);
+    const hash = this._hashMethods.get(kernel);
+
+    if (!hash) {
+      throw new Error(`Kernel (${kernel}) has no hashing params.`);
+    }
+
+    return `${prefix}${hash(code)}${suffix}`;
+  }
+
+  /**
+   * Sets the hash parameters for a kernel.
+   *
+   * @param params - Hashing parameters for a kernel.
+   */
+  setHashParams(params: IDebugger.IConfig.HashParams): void {
+    const { kernel, method, seed } = params;
+    switch (method) {
+      case 'Murmur2':
+        this._hashMethods.set(kernel, code => murmur2(code, seed).toString());
+        break;
+      default:
+        throw new Error(`Hash method (${method}) is not supported.`);
+    }
+  }
+
+  /**
+   * Sets the parameters used by the kernel to create temp files (e.g. cells).
+   *
+   * @param params - Temporary file prefix and suffix for a kernel.
+   */
+  setTmpFileParams(params: IDebugger.IConfig.FileParams): void {
+    const { kernel, prefix, suffix } = params;
+    this._fileParams.set(kernel, { kernel, prefix, suffix });
+  }
+
+  private _fileParams = new Map<string, IDebugger.IConfig.FileParams>();
+  private _hashMethods = new Map<string, (code: string) => string>();
+}
